Widen the child count tree-sitter now returns
tree-sitter 0.27 changed `Node::child_count` to return `u32` where it
returned `usize`. `DecoratedNode::child_count` keeps its `usize` return,
because that is what a caller counting children expects, and the widening
cast loses nothing on any platform whisker builds for.
`named_child_count` still returns `usize` upstream, so it needs nothing.

The two plugin packages outside the workspace kept lockfiles pinning
0.26. They resolve their own dependencies, so nothing here updated them,
and a plugin would have built against one tree-sitter while whisker used
another. Whisker documents that skew as a residual risk rather than a
supported state, so the lockfiles move with the workspace.

+++ crates/whisker-types/src/decorated_node.rs
@@ -80,7 +80,7 @@ impl<'a> DecoratedNode<'a> {
 
     /// Returns the total number of children (named and anonymous)
     pub fn child_count(&self) -> usize {
-        self.node.child_count()
+        self.node.child_count() as usize
     }
 
     /// Returns the child at the given index (named or anonymous)
